Fix variable casting in update rules

+++ update_rules.py
@@ -19,7 +19,7 @@ def anti_hebbian(W, X, sigma, tau1, tau2, l):
     for i in tf.range(k):
         for j in tf.range(n):
             W[i, j].assign(tf.clip_by_value(W[i, j] - X[i, j] * tau1
-                                            * theta(sigma[i], tau1) * theta(tau1, tau2), clip_value_min=tf.cast(-l, tf.float32), clip_value_max=tf.cast(l, tf.float32)))
+                                            * theta(sigma[i], tau1) * theta(tau1, tau2), clip_value_min=tf.cast(-l, tf.int64), clip_value_max=tf.cast(l, tf.int64)))
 
 
 def random_walk(W, X, sigma, tau1, tau2, l):
@@ -27,4 +27,4 @@ def random_walk(W, X, sigma, tau1, tau2, l):
     for i in tf.range(k):
         for j in tf.range(n):
             W[i, j].assign(tf.clip_by_value(W[i, j] + X[i, j] * theta(sigma[i], tau1) * theta(
-                tau1, tau2), clip_value_min=tf.cast(-l, tf.float32), clip_value_max=tf.cast(l, tf.float32)))
+                tau1, tau2), clip_value_min=tf.cast(-l, tf.int64), clip_value_max=tf.cast(l, tf.int64)))
